ipmidetectd: cast time_t to unsigned long before %lu
The client-facing status line printed info->last_received.tv_sec (a
time_t) with %lu.  time_t is not necessarily unsigned long -- e.g. a
64-bit time_t on a 32-bit platform, or a signed time_t -- so the bare
%lu is a format/type mismatch (undefined behavior) that can emit a
wrong value.  Since the ipmidetect client parses this line, a mismatch
could also corrupt the reported last-seen time.  Cast explicitly to
(unsigned long).

+++ ipmidetectd/ipmidetectd.c
@@ -636,7 +636,7 @@ _send_ping_data (void)
       char buf[IPMIDETECTD_BUFLEN];
       int len, n;
 
-      len = snprintf (buf, IPMIDETECTD_BUFLEN, "%s %lu\n", info->hostname, info->last_received.tv_sec);
+      len = snprintf (buf, IPMIDETECTD_BUFLEN, "%s %lu\n", info->hostname, (unsigned long)info->last_received.tv_sec);
       if (len >= IPMIDETECTD_BUFLEN)
         err_exit ("len=%d", len);
